Move halight module code into hlnbtc module
We just added the `htlnbtc` module, the term `halight` is now
deprecated.  However this term is exposed on the REST API.  For now just
move everything we can over to the new name _without_ changing the REST
API.
